plex_youtube_classifier/requirements.txt
python-telegram-bot
openai
yt-dlp
python-dotenv
google-generativeai
